Specify interpreter directly in shebang
When munin invokes this plugin, it does not, evidently, set PATH. We
could require the user to set 'env.PATH' in the plugin config, but
that is not entirely satisfactory because it does not allow a zero-
config deployment. Kick this can down the road by assuming this non-
portable location of the interpreter.